+++ tests/testthat/test-add_overall.tbl_summary.R
@@ -1,3 +1,5 @@
+skip_on_cran()
+
 test_that("add_overall.tbl_summary() works", {
   # typical usage works
   expect_error(