+++ tests/testthat/test-as_hux_table.R
@@ -1,3 +1,4 @@
+skip_on_cran()
 skip_if_not(is_pkg_installed("huxtable", reference_pkg = "gtsummary"))
 
 my_tbl_summary <- trial |>