+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* Correct the order of the columns when the `tbl_summary(by)` variables has ten or more levels: a regression introduced in v2.0.0. (#1877)
+
 # gtsummary 2.0.0
 
 ### New Features