+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Restore functionality of `inline_text.tbl_summary(column)` argument to specify a by level when the by variable is a factor: a regression introduced in v2.0.0. (#1883)
 
+* Correct the order of the columns when the `tbl_summary(by)` variables has ten or more levels: a regression introduced in v2.0.0. (#1877)
+
 # gtsummary 2.0.0
 
 ### New Features